requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
opencv-python
numpy
pandas
ultralytics
langgraph
langchain
sqlalchemy
loguru
pytest
pytest-asyncio
httpx
PyYAML
openai
paho-mqtt
